requirements.txt
fastapi>=0.100
uvicorn>=0.22
sqlalchemy>=1.4
psycopg2-binary>=2.9
python-dotenv>=1.0
jinja2>=3.1
python-multipart>=0.0.6